@@ -4,7 +4,7 @@
           Condition="Exists('$(MSBuildThisFile).user')" />
 
   <PropertyGroup>
-    <LangVersion>8</LangVersion>
+    <LangVersion>9</LangVersion>
     <Nullable>enable</Nullable>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <IsTestProject Condition="$(MSBuildProjectFile.EndsWith('.Tests.csproj', StringComparison.OrdinalIgnoreCase))">true</IsTestProject>
