@@ -3,115 +3,112 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Threading;
 
-static class Program
+try
 {
-    static int Main(string[] args)
-    {
-        try
-        {
-            return Run(new Queue<string>(args));
-        }
-        catch (Exception e)
-        {
-            Console.Error.WriteLine(e.GetBaseException().Message);
-            return 0xbd;
-        }
-    }
+    return Run(new Queue<string>(args));
+}
+catch (Exception e)
+{
+    Console.Error.WriteLine(e.GetBaseException().Message);
+    return 0xbd;
+}
 
-    static int Run(Queue<string> args)
+static int Run(Queue<string> args)
+{
+    switch (args.TryDequeue(out var command) ? command : null)
     {
-        switch (args.TryDequeue(out var command) ? command : null)
-        {
-            case null:
-            {
-                while (true)
-                {
-                    var line = Console.In.ReadLine();
-                    if (line is null)
-                        return 0;
-                    try
-                    {
-                        _ = Run(new Queue<string>(line.Split(' ')));
-                    }
-                    catch (InvalidCommandException e)
-                    {
-                        Console.Error.WriteLine(e.Message);
-                    }
-                }
-            }
-            case "prefix":
-            {
-                var prefix = args.TryDequeue(out var arg) ? arg : "> ";
-                return TransformInput(s => prefix + s);
-            }
-            case "upper":
-                return TransformInput(s => s.ToUpperInvariant());
-            case "lower":
-                return TransformInput(s => s.ToLowerInvariant());
-            case "nop":
-                return 0;
-            case "lorem":
-            {
-                var streams = new[] { Console.Out, Console.Error };
-                var i = 0;
-                // cycle through counts & streams
-                for (var si = 0; args.TryDequeue(out var arg);
-                     si = (si + 1) % streams.Length)
-                {
-                    var stream = streams[si];
-                    var count = int.Parse(arg, NumberStyles.None, CultureInfo.InvariantCulture);
-                    for (; count > 0; count--, i = (i + 1) % LoremIpsums.Length)
-                        stream.WriteLine(LoremIpsums[i]);
-                }
-                return 0;
-            }
-            case "error":
-            {
-                var message = args.TryDequeue(out var arg) && arg.Length > 0
-                            ? arg
-                            : null;
-                throw new ApplicationException(message);
-            }
-            case "exit":
-            {
-                var code = args.TryDequeue(out var arg)
-                         ? int.Parse(arg, NumberStyles.None, CultureInfo.InvariantCulture)
-                         : 0;
-                Environment.Exit(code);
-                return code; // should never get here
-            }
-            case "sleep":
-            {
-                var duration = args.TryDequeue(out var arg)
-                             ? TimeSpan.FromSeconds(double.Parse(arg, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture))
-                             : throw new Exception("Missing seconds argument.");
-                Thread.Sleep(duration);
-                return 0;
-            }
-            default:
-            {
-                throw new InvalidCommandException($"Unknown command: {command}.");
-            }
-        }
-
-        static int TransformInput(Func<string, string> transformer)
+        case null:
         {
             while (true)
             {
                 var line = Console.In.ReadLine();
                 if (line is null)
                     return 0;
-                Console.WriteLine(transformer(line));
+                try
+                {
+                    _ = Run(new Queue<string>(line.Split(' ')));
+                }
+                catch (InvalidCommandException e)
+                {
+                    Console.Error.WriteLine(e.Message);
+                }
             }
+        }
+        case "prefix":
+        {
+            var prefix = args.TryDequeue(out var arg) ? arg : "> ";
+            return TransformInput(s => prefix + s);
+        }
+        case "upper":
+            return TransformInput(s => s.ToUpperInvariant());
+        case "lower":
+            return TransformInput(s => s.ToLowerInvariant());
+        case "nop":
+            return 0;
+        case "lorem":
+        {
+            var streams = new[] { Console.Out, Console.Error };
+            var i = 0;
+            // cycle through counts & streams
+            for (var si = 0; args.TryDequeue(out var arg);
+                 si = (si + 1) % streams.Length)
+            {
+                var stream = streams[si];
+                var count = int.Parse(arg, NumberStyles.None, CultureInfo.InvariantCulture);
+                for (; count > 0; count--, i = (i + 1) % LoremIpsum.Samples.Length)
+                    stream.WriteLine(LoremIpsum.Samples[i]);
+            }
+            return 0;
+        }
+        case "error":
+        {
+            var message = args.TryDequeue(out var arg) && arg.Length > 0
+                        ? arg
+                        : null;
+            throw new ApplicationException(message);
+        }
+        case "exit":
+        {
+            var code = args.TryDequeue(out var arg)
+                     ? int.Parse(arg, NumberStyles.None, CultureInfo.InvariantCulture)
+                     : 0;
+            Environment.Exit(code);
+            return code; // should never get here
+        }
+        case "sleep":
+        {
+            var duration = args.TryDequeue(out var arg)
+                         ? TimeSpan.FromSeconds(double.Parse(arg, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture))
+                         : throw new Exception("Missing seconds argument.");
+            Thread.Sleep(duration);
+            return 0;
+        }
+        default:
+        {
+            throw new InvalidCommandException($"Unknown command: {command}.");
         }
     }
 
-    sealed class InvalidCommandException : Exception
+    static int TransformInput(Func<string, string> transformer)
     {
-        public InvalidCommandException(string message) : base(message) {}
+        while (true)
+        {
+            var line = Console.In.ReadLine();
+            if (line is null)
+                return 0;
+            Console.WriteLine(transformer(line));
+        }
     }
+}
 
-    static readonly string[] LoremIpsums =
+sealed class InvalidCommandException : Exception
+{
+    public InvalidCommandException(string message) : base(message) {}
+}
+
+static class LoremIpsum
+{
+    public static readonly string[] Samples =
     {
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
         "Nullam suscipit nunc non nulla euismod ornare.",
